timob-6092: implement a new activity indicator and rename the old activity indicator as ProgressDialog
