+++ android/modules/ui/src/java/ti/modules/titanium/ui/ActivityIndicatorProxy.java
@@ -10,21 +10,24 @@ import org.appcelerator.kroll.KrollDict;
 import org.appcelerator.kroll.annotations.Kroll;
 import org.appcelerator.titanium.TiC;
 import org.appcelerator.titanium.TiContext;
+import org.appcelerator.titanium.proxy.TiViewProxy;
 import org.appcelerator.titanium.view.TiUIView;
 
 import ti.modules.titanium.ui.widget.TiUIActivityIndicator;
+import ti.modules.titanium.ui.widget.TiUIProgressDialog;
 import android.app.Activity;
 
 @Kroll.proxy(creatableInModule=UIModule.class, propertyAccessors = {
-	"message", "value",
-	"location", "min", "max",
-	"messageid", "type",
-	TiC.PROPERTY_CANCELABLE
+	TiC.PROPERTY_MESSAGE,
+	TiC.PROPERTY_MESSAGEID,
+	TiC.PROPERTY_COLOR,
+	TiC.PROPERTY_FONT,
+	TiC.PROPERTY_STYLE
 })
 @Kroll.dynamicApis(methods = {
 	"hide", "show"
 })
-public class ActivityIndicatorProxy extends TiDialogProxy
+public class ActivityIndicatorProxy extends TiViewProxy
 {
 	public ActivityIndicatorProxy()
 	{
@@ -37,31 +40,30 @@ public class ActivityIndicatorProxy extends TiDialogProxy
 	}
 
 	@Override
-	protected KrollDict getLangConversionTable() {
-		KrollDict table = new KrollDict();
-		table.put("message", "messageid");
-		return table;
-	}
-
-	@Override
 	public TiUIView createView(Activity activity)
 	{
 		return new TiUIActivityIndicator(this);
 	}
 
 	@Override
-	protected void handleShow(KrollDict options) {
+	protected void handleShow(KrollDict options)
+	{
+		if (view == null) {
+			TiUIActivityIndicator ai = (TiUIActivityIndicator) getOrCreateView();
+			ai.show();
+			return;
+		}
 		super.handleShow(options);
-
-		TiUIActivityIndicator ai = (TiUIActivityIndicator) getOrCreateView();
-		ai.show(options);
 	}
 
 	@Override
-	protected void handleHide(KrollDict options) {
+	protected void handleHide(KrollDict options)
+	{
+		if (view == null) {
+			TiUIActivityIndicator ai = (TiUIActivityIndicator) getOrCreateView();
+			ai.hide();
+			return;
+		}
 		super.handleHide(options);
-
-		TiUIActivityIndicator ai = (TiUIActivityIndicator) getOrCreateView();
-		ai.hide(options);
 	}
 }
